+++ Rent/Services/Interfaces/ILandLordService.cs
@@ -8,5 +8,5 @@ public interface ILandLordService
 {
    Task<CommonResponseDto> CreatePropertiesService(propetiesDto data, string useridClaims);
    Task<CommonResponseDto> GetPropertyService(long id);
-   Task<CommonResponseDto>CreateUnitsService(double UserID, unitDto data);
+   Task<CommonResponseDto>CreateUnitsService(long UserID, unitDto data);
 }
